cleaned up login page
