+++ src/action-creators/global-actions.js
@@ -1,0 +1,5 @@
+import { push } from 'react-router-redux';
+
+// Navigation
+export const gotoSignUp = () => push('/sign-up');
+export const gotoHome = () => push('/home');
