--- src/containers/app/New Folder With Items/App.js
@@ -1,21 +1,0 @@
-import React, { Component } from 'react';
-// import logo from './logo.svg';
-// <img src={logo} className="App-logo" alt="logo" />
-import './App.css';
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
-    );
-  }
-}
-
-export default App;
